+++ crates/tor-chanmgr/src/mgr.rs
@@ -15,6 +15,7 @@ use tor_error::internal;
 use tor_linkspec::{HasRelayIds, RelayIds};
 use tor_netdir::params::NetParameters;
 use tor_proto::channel::params::ChannelPaddingInstructionsUpdates;
+use tor_proto::memquota::{ChannelAccount, SpecificAccount as _, ToplevelAccount};
 
 mod select;
 mod state;
@@ -74,6 +75,7 @@ pub(crate) trait AbstractChannelFactory {
         &self,
         target: &Self::BuildSpec,
         reporter: BootstrapReporter,
+        memquota: ChannelAccount,
     ) -> Result<Arc<Self::Channel>>;
 
     /// Construct a new channel for an incoming connection.
@@ -82,6 +84,7 @@ pub(crate) trait AbstractChannelFactory {
         &self,
         peer: std::net::SocketAddr,
         stream: Self::Stream,
+        memquota: ChannelAccount,
     ) -> Result<Arc<Self::Channel>>;
 }
 
@@ -102,6 +105,9 @@ pub(crate) struct AbstractChanMgr<CF: AbstractChannelFactory> {
 
     /// A bootstrap reporter to give out when building channels.
     pub(crate) reporter: BootstrapReporter,
+
+    /// The memory quota account that every channel will be a child of
+    pub(crate) memquota: ToplevelAccount,
 }
 
 /// Type alias for a future that we wait on to see when a pending
@@ -120,10 +126,12 @@ impl<CF: AbstractChannelFactory + Clone> AbstractChanMgr<CF> {
         dormancy: Dormancy,
         netparams: &NetParameters,
         reporter: BootstrapReporter,
+        memquota: ToplevelAccount,
     ) -> Self {
         AbstractChanMgr {
             channels: state::MgrState::new(connector, config.clone(), dormancy, netparams),
             reporter,
+            memquota,
         }
     }
 
@@ -171,8 +179,9 @@ impl<CF: AbstractChannelFactory + Clone> AbstractChanMgr<CF> {
         stream: CF::Stream,
     ) -> Result<Arc<CF::Channel>> {
         let chan_builder = self.channels.builder();
+        let memquota = ChannelAccount::new(&self.memquota)?;
         let _outcome = chan_builder
-            .build_channel_using_incoming(src, stream)
+            .build_channel_using_incoming(src, stream, memquota)
             .await?;
 
         // TODO RELAY: we need to do something with the channel here now that we've created it
@@ -270,8 +279,9 @@ impl<CF: AbstractChannelFactory + Clone> AbstractChanMgr<CF> {
                 // We need to launch a channel.
                 Some(Action::Launch((send, pending_id))) => {
                     let connector = self.channels.builder();
+                    let memquota = ChannelAccount::new(&self.memquota)?;
                     let outcome = connector
-                        .build_channel(&target, self.reporter.clone())
+                        .build_channel(&target, self.reporter.clone(), memquota)
                         .await;
                     let status = self.handle_build_outcome(&target, pending_id, outcome);
 
@@ -651,6 +661,7 @@ mod test {
             Default::default(),
             &Default::default(),
             BootstrapReporter::fake(),
+            ToplevelAccount::new_noop(),
         )
     }
 
@@ -686,6 +697,7 @@ mod test {
             &self,
             target: &Self::BuildSpec,
             _reporter: BootstrapReporter,
+            _memquota: ChannelAccount,
         ) -> Result<Arc<FakeChannel>> {
             yield_now().await;
             let FakeBuildSpec(ident, mood, id) = *target;
@@ -714,6 +726,7 @@ mod test {
             &self,
             _peer: std::net::SocketAddr,
             _stream: Self::Stream,
+            _memquota: ChannelAccount,
         ) -> Result<Arc<Self::Channel>> {
             unimplemented!()
         }